+++ src/components/common/Team/TeamDataCard.module.scss
@@ -1,0 +1,72 @@
+.teamData {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0.8rem 0;
+
+  &:not(:first-child) {
+    border-top: 0.1rem solid $color-gray-200;
+  }
+}
+
+.participants {
+  @include font-body1;
+
+  display: flex;
+  align-items: center;
+  gap: 0.6rem;
+  color: $color-gray-700;
+}
+
+.orderComplete {
+  @include font-body1;
+
+  padding: 0.8rem 1.6rem;
+  color: $color-gray-300;
+}
+
+.creator {
+  @include font-body1;
+
+  color: $color-black;
+}
+
+.timeAndBtn {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 0.8rem;
+}
+
+.timeBox {
+  display: flex;
+  padding: 0.4rem;
+  gap: 0.8rem;
+}
+
+.closed {
+  @include font-body5;
+
+  color: $color-pink-main;
+}
+
+.timer {
+  @include font-body6;
+
+  color: $color-pink-main;
+}
+
+.participationBtn {
+  @include font-headline3;
+
+  display: flex;
+  padding: 0.8rem 1.6rem;
+  justify-content: center;
+  align-items: center;
+  gap: 0.8rem;
+  border-radius: 0.8rem;
+  border: none;
+  background: $color-pink-main;
+  color: $color-white;
+}
